renaming json file
