+++ gen_sig.js
@@ -1,6 +1,6 @@
 const secretKey = '9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08';
 
-const COMMITMENTS = require('./commitments-p256-signed.json');
+const COMMITMENTS = require('./commitments-p256.json');
 
 const bls = require('bls-lib');
